+++ client/src/components/manager/index.js
@@ -1,1 +1,4 @@
-export { default as ManagerHomePage } from './ManagerHomePage';
+export { default as Report } from './Report';
+export { default as StatisticsTable } from './StatisticsTable';
+export { default as CustomBarChart } from './CustomBarChart';
+export { default as CustomPieChart } from './CustomPieChart';
